factor out path related function
