+++ server/src/modes/script/javascript.ts
@@ -3,19 +3,17 @@ import { SymbolInformation, SymbolKind, CompletionItem, Location, SignatureHelp,
 import { LanguageMode } from '../languageModes';
 import { getWordAtText } from '../../utils/strings';
 import { VueDocumentRegions } from '../embeddedSupport';
-import { createUpdater, parseVue, isVue } from './preprocess';
+import { createUpdater, parseVue, isVue, getFileFsPath, getFilePath } from './preprocess';
 
 import Uri from 'vscode-uri';
 import * as path from 'path';
 import * as ts from 'typescript';
 import * as _ from 'lodash';
-import { platform } from 'os';
 
 import { NULL_SIGNATURE, NULL_COMPLETION } from '../nullMode';
 
 import * as bridge from './bridge';
 
-const IS_WINDOWS = platform() === 'win32';
 const JS_WORD_REGEX = /(-?\d*\.\d\w*)|([^\`\~\!\@\#\%\^\&\*\(\)\-\=\+\[\{\]\}\\\|\;\:\'\"\,\.\<\>\/\?\s]+)/g;
 
 export interface ScriptMode extends LanguageMode {
@@ -460,19 +458,6 @@ export function getJavascriptMode (documentRegions: LanguageModelCache<VueDocume
 
 function getNormalizedFileFsPath (fileName: string): string {
   return Uri.file(fileName).fsPath;
-}
-
-function getFileFsPath (documentUri: string): string {
-  return Uri.parse(documentUri).fsPath;
-}
-
-function getFilePath (documentUri: string): string {
-  if (IS_WINDOWS) {
-    // Windows have a leading slash like /C:/Users/pine
-    return Uri.parse(documentUri).path.slice(1);
-  } else {
-    return Uri.parse(documentUri).path;
-  }
 }
 
 function languageServiceIncludesFile (ls: ts.LanguageService, documentUri: string): boolean {
